+++ core/Updates/0.2.13.php
@@ -9,30 +9,39 @@
 
 namespace Piwik\Updates;
 
-use Piwik\Common;
 use Piwik\Updater;
 use Piwik\Updates;
+use Piwik\Updater\Migration\Factory as MigrationFactory;
 
 /**
  */
 class Updates_0_2_13 extends Updates
 {
-    public function getMigrationQueries(Updater $updater)
+    /**
+     * @var MigrationFactory
+     */
+    private $migration;
+
+    public function __construct(MigrationFactory $factory)
+    {
+        $this->migration = $factory;
+    }
+
+    public function getMigrations(Updater $updater)
     {
         return array(
-            'DROP TABLE IF EXISTS `' . Common::prefixTable('option') . '`'    => false,
-
-            'CREATE TABLE `' . Common::prefixTable('option') . "` (
-				option_name VARCHAR( 64 ) NOT NULL ,
-				option_value LONGTEXT NOT NULL ,
-				autoload TINYINT NOT NULL DEFAULT '1',
-				PRIMARY KEY ( option_name )
-			)" => 1050,
+            $this->migration->db->dropTable('option'),
+            $this->migration->db->createTable('option', array(
+                'option_name'  => 'VARCHAR( 64 ) NOT NULL' ,
+                'option_value' => 'LONGTEXT NOT NULL' ,
+                'autoload' => "TINYINT NOT NULL DEFAULT '1'",
+            )),
+            $this->migration->db->addPrimaryKey('option', 'option_name')
         );
     }
 
     public function doUpdate(Updater $updater)
     {
-        $updater->executeMigrationQueries(__FILE__, $this->getMigrationQueries($updater));
+        $updater->executeMigrations(__FILE__, $this->getMigrations($updater));
     }
 }