+++ core/Updates/1.5-b2.php
@@ -9,33 +9,48 @@
 
 namespace Piwik\Updates;
 
-use Piwik\Common;
 use Piwik\Updater;
 use Piwik\Updates;
+use Piwik\Updater\Migration\Factory as MigrationFactory;
 
 /**
  */
 class Updates_1_5_b2 extends Updates
 {
-    public function getMigrationQueries(Updater $updater)
+    /**
+     * @var MigrationFactory
+     */
+    private $migration;
+
+    public function __construct(MigrationFactory $factory)
     {
+        $this->migration = $factory;
+    }
+
+    public function getMigrations(Updater $updater)
+    {
+        $customVarType = 'VARCHAR(100) DEFAULT NULL';
+
+        $customVarColumns = array(
+            'custom_var_k1' => $customVarType,
+            'custom_var_v1' => $customVarType,
+            'custom_var_k2' => $customVarType,
+            'custom_var_v2' => $customVarType,
+            'custom_var_k3' => $customVarType,
+            'custom_var_v3' => $customVarType,
+            'custom_var_k4' => $customVarType,
+            'custom_var_v4' => $customVarType,
+            'custom_var_k5' => $customVarType,
+            'custom_var_v5' => $customVarType,
+        );
+
         return array(
-            'ALTER TABLE `' . Common::prefixTable('log_link_visit_action') . '`
-				 ADD  custom_var_k1 VARCHAR(100) DEFAULT NULL AFTER time_spent_ref_action,
-				 ADD  custom_var_v1 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_k2 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_v2 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_k3 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_v3 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_k4 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_v4 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_k5 VARCHAR(100) DEFAULT NULL,
-				 ADD  custom_var_v5 VARCHAR(100) DEFAULT NULL' => 1060,
+            $this->migration->db->addColumns('log_link_visit_action', $customVarColumns, 'time_spent_ref_action')
         );
     }
 
     public function doUpdate(Updater $updater)
     {
-        $updater->executeMigrationQueries(__FILE__, $this->getMigrationQueries($updater));
+        $updater->executeMigrations(__FILE__, $this->getMigrations($updater));
     }
 }