+++ core/Updates/3.0.0-b1.php
@@ -17,6 +17,8 @@ use Piwik\DbHelper;
 use Piwik\Option;
 use Piwik\Plugins\Installation\ServerFilesGenerator;
 use Piwik\Updater;
+use Piwik\Updater\Migration;
+use Piwik\Updater\Migration\Factory as MigrationFactory;
 use Piwik\Updates;
 use Piwik\Plugins\Dashboard;
 
@@ -26,20 +28,34 @@ use Piwik\Plugins\Dashboard;
 class Updates_3_0_0_b1 extends Updates
 {
     /**
-     * Here you can define one or multiple SQL statements that should be executed during the update.
-     * @return array
+     * @var MigrationFactory
      */
-    public function getMigrationQueries(Updater $updater)
+    private $migration;
+
+    private $pluginSettingsTable = 'plugin_setting';
+    private $siteSettingsTable = 'site_setting';
+
+    public function __construct(MigrationFactory $factory)
+    {
+        $this->migration = $factory;
+    }
+
+    /**
+     * Here you can define one or multiple SQL statements that should be executed during the update.
+     * @param Updater $updater
+     * @return Migration[]
+     */
+    public function getMigrations(Updater $updater)
     {
         $db = Db::get();
         $allGoals = $db->fetchAll(sprintf("SELECT DISTINCT idgoal FROM %s", Common::prefixTable('goal')));
         $allDashboards = $db->fetchAll(sprintf("SELECT * FROM %s", Common::prefixTable('user_dashboard')));
 
-        $queries = $this->getDashboardMigrationSqls($allDashboards, $allGoals);
-        $queries = $this->getPluginSettingsMigrationQueries($queries, $db);
-        $queries = $this->getSiteSettingsMigrationQueries($queries);
+        $migrations = $this->getDashboardMigrations($allDashboards, $allGoals);
+        $migrations = $this->getPluginSettingsMigrations($migrations);
+        $migrations = $this->getSiteSettingsMigrations($migrations);
 
-        return $queries;
+        return $migrations;
     }
 
     public function doUpdate(Updater $updater)
@@ -64,25 +80,18 @@ class Updates_3_0_0_b1 extends Updates
     }
 
     /**
-     * @param $queries
-     * @param Db $db
-     * @return mixed
+     * @param Migration[] $queries
+     * @return Migration[]
      */
-    private function getPluginSettingsMigrationQueries($queries, $db)
+    private function getPluginSettingsMigrations($queries)
     {
-        $pluginSettingsTableName = $this->getPluginSettingsTableName();
-        $dbSettings = new Db\Settings();
-        $engine = $dbSettings->getEngine();
-
-        $pluginSettingsTable = "CREATE TABLE $pluginSettingsTableName (
-                          `plugin_name` VARCHAR(60) NOT NULL,
-                          `setting_name` VARCHAR(255) NOT NULL,
-                          `setting_value` LONGTEXT NOT NULL,
-                          `user_login` VARCHAR(100) NOT NULL DEFAULT '',
-                              INDEX(plugin_name, user_login)
-                            ) ENGINE=$engine DEFAULT CHARSET=utf8
-            ";
-        $queries[$pluginSettingsTable] = 1050;
+        $queries[] = $this->migration->db->createTable($this->pluginSettingsTable, array(
+            'plugin_name' => 'VARCHAR(60) NOT NULL',
+            'setting_name' => 'VARCHAR(255) NOT NULL',
+            'setting_value' => 'LONGTEXT NOT NULL',
+            'user_login' => "VARCHAR(100) NOT NULL DEFAULT ''",
+        ));
+        $queries[] = $this->migration->db->addIndex($this->pluginSettingsTable, array('plugin_name', 'user_login'));
 
         $optionTable = Common::prefixTable('option');
         $query = 'SELECT `option_name`, `option_value` FROM `' . $optionTable . '` WHERE `option_name` LIKE "Plugin_%_Settings"';
@@ -103,61 +112,54 @@ class Updates_3_0_0_b1 extends Updates
                 }
 
                 foreach ($settingValue as $val) {
-                    $queries[$this->createPluginSettingQuery($pluginName, $settingName, $val)] = 1062;
+                    $queries[] = $this->createPluginSettingQuery($pluginName, $settingName, $val);
                 }
             }
         }
 
-        $queries[$query = sprintf('DELETE FROM `%s` WHERE `option_name` like "Plugin_%%_Settings"', $optionTable)] = false;
+        $queries[] = $this->migration->db->sql(sprintf('DELETE FROM `%s` WHERE `option_name` like "Plugin_%%_Settings"', $optionTable));
 
         return $queries;
     }
 
     /**
-     * @param $queries
-     * @param Db $db
-     * @return mixed
+     * @param Migration[] $queries
+     * @return Migration[]
      */
-    private function getSiteSettingsMigrationQueries($queries)
+    private function getSiteSettingsMigrations($queries)
     {
-        $table = Common::prefixTable('site_setting');
+        $table = $this->siteSettingsTable;
+        $queries[] = $this->migration->db->addColumn($table, 'plugin_name', 'VARCHAR(60) NOT NULL', $afer = 'idsite');
 
-        $pluginSettingsTable = "ALTER TABLE $table ADD COLUMN `plugin_name` VARCHAR(60) NOT NULL AFTER `idsite`";
-        $queries[$pluginSettingsTable] = 1060;
-        $queries["ALTER TABLE $table DROP PRIMARY KEY, ADD INDEX(idsite, plugin_name);"] = false;
+        $table = Common::prefixTable($table);
+        $queries[] = $this->migration->db->sql("ALTER TABLE `$table` DROP PRIMARY KEY, ADD INDEX(idsite, plugin_name);",
+                                               Migration\Db::ERROR_CODE_COLUMN_NOT_EXISTS);
 
         // we cannot migrate existing settings as we do not know the related plugin name, but this feature
         // (measurablesettings) was not really used anyway. If a migration is somewhere really needed it has to be
         // handled in the plugin
-        $queries[sprintf('DELETE FROM `%s`', $table)] = false;
+        $queries[] = $this->migration->db->sql(sprintf('DELETE FROM `%s`', $table));
 
         return $queries;
     }
 
     private function createPluginSettingQuery($pluginName, $settingName, $settingValue)
     {
-        $table = $this->getPluginSettingsTableName();
-
         $login = '';
         if (preg_match('/^.+#(.+)#$/', $settingName, $matches)) {
             $login = $matches[1];
             $settingName = str_replace('#' . $login . '#', '', $settingName);
         }
 
-        $db = Db::get();
-
-        $query  = sprintf("INSERT INTO %s (plugin_name, setting_name, setting_value, user_login) VALUES ", $table);
-        $query .= sprintf("(%s, %s, %s, %s)", $db->quote($pluginName), $db->quote($settingName), $db->quote($settingValue), $db->quote($login));
-
-        return $query;
+        return $this->migration->db->insert($this->pluginSettingsTable, array(
+            'plugin_name' => $pluginName,
+            'setting_name' => $settingName,
+            'setting_value' => $settingValue,
+            'user_login' => $login
+        ));
     }
 
-    private function getPluginSettingsTableName()
-    {
-        return Common::prefixTable('plugin_setting');
-    }
-
-    private function getDashboardMigrationSqls($allDashboards, $allGoals)
+    private function getDashboardMigrations($allDashboards, $allGoals)
     {
         $sqls = array();
 
@@ -444,6 +446,9 @@ class Updates_3_0_0_b1 extends Updates
                     ));
         }
 
+        $table = Common::prefixTable('user_dashboard');
+        $sql = sprintf('UPDATE %s SET layout = ? WHERE iddashboard = ?', $table);
+
         foreach ($allDashboards as $dashboard) {
             $dashboardLayout = json_decode($dashboard['layout']);
 
@@ -451,7 +456,7 @@ class Updates_3_0_0_b1 extends Updates
 
             $newLayout = json_encode($dashboardLayout);
             if ($newLayout != $dashboard['layout']) {
-                $sqls["UPDATE " . Common::prefixTable('user_dashboard') . " SET layout = '".addslashes($newLayout)."' WHERE iddashboard = ".$dashboard['iddashboard']] = false;
+                $sqls[] = $this->migration->db->boundSql($sql, array($newLayout, $dashboard['iddashboard']));
             }
         }