+++ CHANGELOG.md
@@ -26,6 +26,10 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 * The creation of settings has slightly changed to improve performance. It is now possible to create new settings via the method `$this->makeSetting()` see `Piwik\Plugins\ExampleSettingsPlugin\SystemSettings` for an example.
 * It is no possible to define an introduction text for settings.
 
+### Deprecations
+* The method `Piwik\Updates::getMigrationQueries()` has been deprecated and renamed to `getMigrations()`. It is still supported to use the method, but the method will be removed in Piwik 4.0.0
+* The method `Piwik\Updater::executeMigrationQueries()` has been deprecated and renamed to `executeMigrations`. It is still supported to use the method, but the method will be removed in Piwik 4.0.0.
+
 ### New APIs
 * Multiple widgets for one report can now be created via the `Report::configureWidgets()` method via the new classes `Piwik\Widget\ReportWidgetFactory` and `Piwik\Widget\ReportWidgetConfig`
 * There is a new property `Report::$subCategory` that let's you add a report to the reporting UI. If a page having that name does not exist yet, it will be created automatically. The newly added method `Report::getSubCategory()` let's you get this value.
@@ -35,6 +39,8 @@ This is a changelog for Piwik platform developers. All changes for our HTTP API'
 * New HTTP API method `API.getReportPagesMetadata` to get a list of all available pages that exist including the widgets they include
 * New HTTP API method `SitesManager.getSiteSettings` to get a list of all available settings for a specific site
 * The JavaScript AjaxHelper has a new method `ajaxHelper.withTokenInUrl()` to easily send a token along a XHR. Within the Controller the existence of this token can be checked via `$this->checkTokenInUrl();` to prevent CSRF attacks.
+* The new class `Piwik\Updater\Migration\Factory` lets you easily create migrations that can be executed during an update. For example database or plugin related migrations. To generate a new update with migrations execute `./console generate:update`.
+* The new method `Piwik\Updater::executeMigration` lets you execute a single migration.
 
 ### New features
 * New "Sparklines" visualization that let's you create a widget showing multiple sparklines